avoid memory errors on copying data to other state
Attempts to avoid the following possible unhandled errors in scenarios like:

```C
lua_pushstring(L2, lua_tostring(L1, -1));
```

If this code is running in `L1` and `L2` is suspended. If there is an memory
error in `lua_pushstring`, this would make the execution "jump" to the
`lua_pcall` of `L2`, which is the suspended thread.

The suggested approach is to avoid this copying direction (from running to
suspended), and only copy in the opposite diection (from suspended to running).